+++ src/main/java/io/github/astrapi69/jaxb/ClassToXsdExtensions.java
@@ -24,23 +24,27 @@
  */
 package io.github.astrapi69.jaxb;
 
+import java.io.File;
 import java.io.IOException;
 
+import io.github.astrapi69.file.write.WriteFileExtensions;
 import io.github.astrapi69.jaxb.schema.XsdSchemaOutputResolver;
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBException;
 
 /**
  * The class {@link ClassToXsdExtensions} provides methods for convert java class objects to xsd
- * string objects
+ * string objects or to xsd file. <br/>
+ * For transform them back you java classes, you have to use xjc. <br/>
+ * <br/>
+ * An example is provided in the example repository
+ * <a href="https://github.com/astrapi69/jaxb-exjc">jaxb-exjc</a>
  */
 public class ClassToXsdExtensions
 {
 
 	/**
-	 * Converts the given java class objects to xsd string that can be used in an xsd file. For
-	 * transform them back you java classes you have to use xjc. An example is provided in the
-	 * example repository <a href="https://github.com/astrapi69/jaxb-exjc">jaxb-exjc</a>
+	 * Converts the given java class objects to a xsd string
 	 *
 	 * @param classesToBeBound
 	 *            the java class objects
@@ -50,11 +54,71 @@ public class ClassToXsdExtensions
 	 * @throws IOException
 	 *             if an error occur when the {@link JAXBContext} try to generate the schema
 	 */
-	public static String classToXsd(Class<?>... classesToBeBound) throws JAXBException, IOException
+	public static String classesToXsdString(Class<?>... classesToBeBound)
+		throws JAXBException, IOException
 	{
 		JAXBContext jaxbContext = JAXBContext.newInstance(classesToBeBound);
 		XsdSchemaOutputResolver schemaOutputResolver = new XsdSchemaOutputResolver();
 		jaxbContext.generateSchema(schemaOutputResolver);
 		return schemaOutputResolver.getResult();
 	}
+
+	/**
+	 * Converts the given java class object to a xsd string
+	 *
+	 * @param classToBeBound
+	 *            the java class object
+	 * @return the xsd string
+	 * @throws JAXBException
+	 *             if an error occurs by creating a new instance of the {@link JAXBContext}
+	 * @throws IOException
+	 *             if an error occur when the {@link JAXBContext} try to generate the schema
+	 */
+	public static <T> String classToXsdString(Class<T> classToBeBound)
+		throws JAXBException, IOException
+	{
+		JAXBContext jaxbContext = JAXBContext.newInstance(classToBeBound);
+		XsdSchemaOutputResolver schemaOutputResolver = new XsdSchemaOutputResolver();
+		jaxbContext.generateSchema(schemaOutputResolver);
+		return schemaOutputResolver.getResult();
+	}
+
+	/**
+	 * Converts the given java class object into the given xsd file
+	 *
+	 * @param classToBeBound
+	 *            the java class object
+	 * @param xsdOutputFile
+	 *            the xsd file to write the result
+	 * @throws JAXBException
+	 *             if an error occurs by creating a new instance of the {@link JAXBContext}
+	 * @throws IOException
+	 *             if an error occur when the {@link JAXBContext} try to generate the schema
+	 */
+	public static <T> void classToXsdFile(Class<T> classToBeBound, File xsdOutputFile)
+		throws JAXBException, IOException
+	{
+		WriteFileExtensions.writeStringToFile(xsdOutputFile, classToXsdString(classToBeBound),
+			"UTF-8");
+	}
+
+	/**
+	 * Converts the given java class objects into the given xsd file
+	 *
+	 * @param xsdOutputFile
+	 *            the xsd file to write the result
+	 * @param classesToBeBound
+	 *            the java class objects
+	 * @throws JAXBException
+	 *             if an error occurs by creating a new instance of the {@link JAXBContext}
+	 * @throws IOException
+	 *             if an error occur when the {@link JAXBContext} try to generate the schema
+	 */
+	public static void classesToXsdFile(File xsdOutputFile, Class<?>... classesToBeBound)
+		throws JAXBException, IOException
+	{
+		WriteFileExtensions.writeStringToFile(xsdOutputFile, classesToXsdString(classesToBeBound),
+			"UTF-8");
+	}
+
 }