+++ src/main/java/module-info.java
@@ -25,6 +25,7 @@ module jaxb.extensions.main
 	requires throwable;
 	requires xml.api.main;
 	requires silly.io.main;
+	requires file.worker;
 
 	exports io.github.astrapi69.jaxb;
 	exports io.github.astrapi69.jaxb.factory;